Move the SiteWide Notice block into the Members component
