+++ src/bp-core/bp-core-cssjs.php
@@ -81,6 +81,10 @@ function bp_core_register_common_scripts() {
 	$version = bp_get_version();
 	foreach ( $scripts as $id => $script ) {
 		wp_register_script( $id, $script['file'], $script['dependencies'], $version, $script['footer'] );
+
+		if ( isset( $script['localize']['name'], $script['localize']['data'] ) ) {
+			wp_localize_script( $id, $script['localize']['name'], $script['localize']['data'] );
+		}
 	}
 }
 add_action( 'bp_enqueue_scripts',       'bp_core_register_common_scripts', 1 );
