+++ src/bp-members/blocks/close-notices-block/index.asset.php
