+++ src/bp-members/blocks/close-notices-block/index.js
