+++ src/bp-members/blocks/sitewide-notices/block.json
