+++ src/bp-members/blocks/sitewide-notices/index.asset.php
