+++ src/js/blocks/bp-members/close-notices-block/sitewide-notices.js
