+++ src/js/blocks/bp-members/sitewide-notices/.babelrc
