+++ src/js/blocks/bp-members/sitewide-notices/block.json
