+++ src/js/blocks/bp-members/sitewide-notices/sitewide-notices.js
